Adds top level functionality for integration
This commit integrates the stage runner functionality and the
light stemcell building process implemented by the previous commit.

The testing for this level involves a full beginning-to-end process of
extracting the heavy stemcell, creating the AMI for this heavy stemcell,
copying the AMIs to different regions, and then deleting the AMIs.

[#107775580](https://www.pivotaltracker.com/story/show/107775580)

Signed-off-by: Edwin Xie <exie@pivotal.io>