Stemcell builder now outputs the name of the light stemcell
If the stemcell is to be built with HVM, it will insert hvm in the name.
Also, the name of the light stemcell tarball will be `light-*.tgz`. The
input for the output path is now a folder in which the light stemcell
will be placed.

Additionally in this commit:
  - Fixed a typo in the status waiter test