Publish light stemcell for xenial v40
This change updates the aws light stemcell pipeline to publish a new
light stemcell for the v40 of the xenial stemcell.

As a result of having two ubuntu distributions, this change also renames
`ubuntu` to `ubuntu-trusty` and `centos` to `centos-7`. This change also
reorganizes the pipeline groups to be by distribution, rather than
version.

This change also removes unsupported versions of light stemcells.
Namely the alpha builds, 3312 for ubuntu-trusty, 3469 for ubuntu-trusty,
and 3468 for centos-7.

[finishes #157755816](https://www.pivotaltracker.com/story/show/157755816)

Co-authored-by: Difan Zhao <dzhao@pivotal.io>