Use S3 URLs instead of presigned URLs
The ImportSnapshot API doesn't accept presigned URLs that are generated by the
SDK when we use IAM profiles.  I don't think it has anything to do with the way
we are generating the presigned URL.  The output of the following command is
also not accepted by the ImportSnapshot API:

  aws s3 presign s3://bucket/key

It is worth noting that presigned URLs generated when IAM profiles are used
have an extra query parameter `X-Amz-Security-Token'.

Resources:

- https://github.com/aws/aws-sdk-go/issues/1842
- https://docs.aws.amazon.com/general/latest/gr/sigv4-create-canonical-request.html

Authored-by: John Shahid <jshahid@pivotal.io>