Add tagging to created / copied stemcells #182476264
currently there is no metadata whatsoever on the light
stemcells that the builder produces. this creates an
issue when we eventually need to cleanup the account
to remove unpublished build artifacts. this adds tagging
in the form of:
```
published	false/true
version	621.123
distro	ubuntu-xenial
name	ubuntu-xenial-621.123
```

so it is easier to cleanup the unneeded artifacts.

additionally fix some linting issues by removing the
unnecessary type for elements within a typed array.