fix wrong published tag value
Currently creation the stemcells are not published yet. They are
publicly accessible (which is what the code used to determine if they're
published) but that doesn't mean they're actually in TanzuNet or on
bosh.io. Set the published tag to false on create/copy. There is a task
in CI that will update the tags once the publish-ubuntu-* jobs are
running.