Add tagging for snapshots created in AMI creation
If we clean up AMIs AWS doesn't automatically clean the orphaned and not
required snapshots anymore. Currently snapshots that back an AMI are
hard to discover and cleaning takes a long time because snapshots need
to be associated to an ami first (via the AMI data in the description).
Using the description makes this process harder than it needs to be.
This adds tags to a snapshot to associate it with an AMI for later
discoverability.

[ #183700416 ]