Changes in aws-sdk-go v1.44.221 cause the TokenProvider to output error messages when it fails to fetch instance profile credentials.
This output went to stdout by default and was breaking our tests which parse
stdout looking for the metadata.

The output now goes to stderr and can be very noisy. Silencing the logger
seems dangerous as we could miss important errors then. Unfortunately there
is no granularity of log messages or any way to silence this single error.