Revert the changes to prevent creds package from logging to stdout and instead switch on auth type
Rather than always trying to use instance profiles and access keys in a chain, just always
use access keys if they are present, and fall back to instance profiles if they are not.

This prevents verbose new warnings done by the instance profile creds source warning that it cannot
fetch the token.