Fix integration tests by restoring default behavior
Previous commit to prevent publishing ebs snapshots of private AMIs
changed the default behavior of always publishing all the snapshots.

This commit flips the logic in an 'if' test added to prevent publishing
snapshots of private AMIs so that it publishes snapshots UNLESS the AMI
is a private one. (This change lets us leave the logic in the test that
predates the relevant change intact, and preserves the intended
behavior.)

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>