Adds env var to skip replication tests
This commit adds an off-by-default mechanism to skip the AWS replication
tests added in support of the FIPS stemcell work.

We're making this change because we still need to provide new AWS Xenial
stemcells to our customers, but we don't expect there will ever be a
FIPS stemcell, so it doesn't seem all that useful to rework our Xenial
publishing pipeline to run these new tests.

Signed-off-by: Kenneth Lakin <klakin@vmware.com>