Loudly error when failing to make an AMI public
We published a light stemcell (Jammy 1.406) whose AMI in region
`eu-central-1` was not public, causing an error when a user tried to
deploy.

The problem was we never checked to make sure that our attempt to make
the AMI public succeeded.

With this commit, we now propagate any errors that occur when attempting
to make the AMI we publish public.

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>

[#184045754] Ensure the aws light stemcell builder catches errors when making AMIs public